+++ src/add_book.h
@@ -1,0 +1,3 @@
+#pragma once
+#include <sqlite3.h>
+void add_book(sqlite3 *db);
