+++ src/list_book.h
@@ -1,0 +1,2 @@
+#pragma once
+void list_books(sqlite3 *db);
