+++ src/print_help.c
@@ -1,0 +1,9 @@
+#include <stdio.h>
+void print_help() {
+    printf("使用可能なコマンド:\n");
+    printf("  add   - 本を追加\n");
+    printf("  list  - 登録された本の一覧を表示\n");
+    printf("  progress - 読書進捗を記録・表示\n");
+    printf("  help  - ヘルプを表示\n");
+    printf("  exit  - プログラムを終了\n");
+}
