+++ src/print_help.h
@@ -1,0 +1,1 @@
+void print_help();
