+++ src/update_progress.h
@@ -1,0 +1,2 @@
+#pragma once
+void update_progress(sqlite3 *db);
